Refactor logic for determining work deposit pathway
