+++ app/controllers/dashboard/form/work_version_details_controller.rb
@@ -8,15 +8,13 @@ module Dashboard
       end
 
       def edit
-        wv = WorkVersion.includes(:work).find(params[:id])
-        authorize(wv)
-        @resource = WorkDepositPathways.details_form_for(wv)
+        authorize(work_version)
+        @resource = deposit_pathway.details_form
       end
 
       def update
-        wv = WorkVersion.includes(:work).find(params[:id])
-        authorize(wv)
-        @resource = WorkDepositPathways.details_form_for(wv)
+        authorize(work_version)
+        @resource = deposit_pathway.details_form
         @resource.attributes = work_version_params
         process_response(on_error: :edit)
       end
@@ -47,6 +45,14 @@ module Dashboard
 
         def next_page_path
           dashboard_form_contributors_path('work_version', @resource.id)
+        end
+
+        def work_version
+          @work_version ||= WorkVersion.includes(:work).find(params[:id])
+        end
+
+        def deposit_pathway
+          @deposit_pathway ||= WorkDepositPathway.new(work_version)
         end
     end
   end
