+++ spec/models/work_spec.rb
@@ -631,59 +631,4 @@ RSpec.describe Work, type: :model do
       expect(work.latest_published_version_dois).to eq([work_version.doi, work_version.identifier].flatten.map { |n| "doi:#{n}" })
     end
   end
-
-  describe '#deposit_pathway' do
-    %w[
-      article
-      book
-      capstone_project
-      conference_proceeding
-      dissertation
-      masters_culminating_experience
-      masters_thesis
-      part_of_book
-      report
-      research_paper
-      thesis
-    ].each do |t|
-      context "when the work's type is '#{t}'" do
-        it 'returns :scholarly_works' do
-          w = described_class.new(work_type: t)
-          expect(w.deposit_pathway).to eq :scholarly_works
-        end
-      end
-    end
-
-    %w[
-      dataset
-      software_or_program_code
-    ].each do |t|
-      context "when the work's type is '#{t}'" do
-        it 'returns :data_and_code' do
-          w = described_class.new(work_type: t)
-          expect(w.deposit_pathway).to eq :data_and_code
-        end
-      end
-    end
-
-    %w[
-      audio
-      image
-      journal
-      map_or_cartographic_material
-      other
-      poster
-      presentation
-      project
-      unspecified
-      video
-    ].each do |t|
-      context "when the work's type is '#{t}'" do
-        it 'returns :general' do
-          w = described_class.new(work_type: t)
-          expect(w.deposit_pathway).to eq :general
-        end
-      end
-    end
-  end
 end
